renamed Site.partner to Site.is_partner Bugfixes with API-functions async update_sites()

+++ kinkcom/views.py
@@ -11,6 +11,7 @@ from kinkcom.models import KinkComSite, KinkComShoot, KinkComPerformer
 
 
 def shoot(request, shootid=None, title=None, date=None, performer_number=None, performer_name=None):
+    return_dict = {'errors':None, 'length':0}
     if shootid:
         shoots_ = _get_shoots_by_shootid(shootid)
     elif title:
@@ -18,12 +19,13 @@ def shoot(request, shootid=None, title=None, date=None, performer_number=None, p
     elif date:
         try:
             if date.isdigit():
-                date_obj = datetime.date.fromtimestamp(date)
+                date_obj = datetime.date.fromtimestamp(int(date))
             else:
-                date_obj = datetime.datetime.strptime(date, '%B %d, %Y').date()
+                date_obj = datetime.datetime.strptime(date, '%Y-%m-%d').date()
             shoots_ = _get_shoots_by_date(date_obj)
         except ValueError:
-            shoots_ = {'error': 'Cannot recognize date format, must be "%B %d, %Y", was "{}"'.format(date)}
+            return_dict['error'] = 'Cannot recognize date format, must be %Y-%m-%d or %s, was "{}"'.format(date)
+            shoots_ = KinkComShoot.objects.none()
     elif performer_number:
         shoots_ = _get_shoots_by_performer_number(performer_number)
     elif performer_name:
@@ -32,20 +34,25 @@ def shoot(request, shootid=None, title=None, date=None, performer_number=None, p
         shoots_ = KinkComShoot.objects.none()
 
     shoots_ = shoots_ if shoots_ else []
-    j_ = json.dumps(shoots_.serialize())
+    return_dict['results'] = [s.serialize()for s in shoots_]
+    return_dict['length'] = shoots_.count()
+    j_ = json.dumps(return_dict)
     return HttpResponse(j_, content_type = 'application/json; charset=utf8')
 
 
 def performer(request, performer_name=None, performer_number=None):
+    return_dict = {'errors': None, 'length': 0}
     if performer_number:
-        performer_ = _get_performers_by_number(performer_number)
+        performers_ = _get_performers_by_number(performer_number)
     elif performer_name:
-        performer_ = _get_performers_by_name(performer_name)
+        performers_ = _get_performers_by_name(performer_name)
     else:
-        performer_ = KinkComPerformer.objects.none()
+        performers_ = KinkComPerformer.objects.none()
 
-    performer_ = performer_ if performer_ is not None else []
-    j_ = json.dumps(performer_.serialize())
+    performers_ = performers_ if performers_ is not None else []
+    return_dict['results'] = [s.serialize() for s in performers_]
+    return_dict['length'] = performers_.count()
+    j_ = json.dumps(return_dict)
     return HttpResponse(j_, content_type = 'application/json; charset=utf8')
 
 
@@ -65,14 +72,10 @@ def _get_performers_by_name(performer_name):
 
 
 def _get_shoots_by_shootid(shootid):
-    try:
-        return KinkComShoot.objects.get(shootid=shootid)
-    except ObjectDoesNotExist:
-        return KinkComShoot.objects.none()
-    except MultipleObjectsReturned:
-        shoots = KinkComShoot.objects.filter(shootid=shootid)
+    shoots = KinkComShoot.objects.filter(shootid=shootid)
+    if shoots.count() > 1:
         [s.delete() for s in shoots[1:]]
-        return shoots[0]
+    return shoots
 
 
 def _get_shoots_by_title(title):
@@ -84,12 +87,14 @@ def _get_shoots_by_date(date_obj):
 
 
 def _get_shoots_by_performer_number(performer_number):
-    return KinkComShoot.objects.filter(performer__number=performer_number)
+    return KinkComShoot.objects.filter(performers__number=performer_number)
 
 
 def _get_shoots_by_performer_name(performer_name):
     shoots_ = Q()
     performers_ = _get_performers_by_name(performer_name)
     for performer_ in performers_:
-        shoots_ |= _get_shoots_by_performer_number(performer_.number)
-    return shoots_
+        shoots_ |= Q(_get_shoots_by_performer_number(performer_.number))
+    if not shoots_:
+        return KinkComShoot.objects.none()
+    return KinkComShoot.objects.filter(shoots_)
